Move more JKSet-dependent functions to temporary deprecated submodule
